fix(test): move isdir monkeypatch inline, revert conftest.py to original
Per review: the global autouse fixture in conftest.py had too broad a
blast radius. Moved the os.path.isdir stub inline into the one test that
needs it, with selective patching for /mnt and /srv paths only.
Reverted conftest.py to its original state—it was not in the brief's
file list and should not appear in this task's diff.
